+++ constrain/Classes/Extensions.swift
@@ -7,18 +7,6 @@
 
 import UIKit
 
-extension UIView {
-    /// Returns a Constraint instance which eases creation of NSLayoutConstraints.
-    /// Save a reference to the Constraint instance for later access to the layout constraints.
-    public var constrain: Constraints {
-        return Constraints(view: self)
-    }
-    
-    public func constrain(withName name: String) -> Constraints {
-        return Constraints(view: self, name: name)
-    }
-}
-
 public extension UIViewController {
     @discardableResult
     func constrainSubview(_ viewController: UIViewController) -> Constraints {
@@ -97,23 +85,4 @@ public extension UIView {
         return constrainSibling(sibling).top(to: self.bottomAnchor, constant: constant, by: relationship, priority: priority)
     }
     
-}
-
-// Safe areas
-extension UIView {
-    public var topAnchorSafe: NSLayoutYAxisAnchor {
-        if #available(iOS 11.0, *) {
-            return safeAreaLayoutGuide.topAnchor
-        } else {
-            return topAnchor
-        }
-    }
-    
-    public var bottomAnchorSafe: NSLayoutYAxisAnchor {
-        if #available(iOS 11.0, *) {
-            return safeAreaLayoutGuide.bottomAnchor
-        } else {
-            return bottomAnchor
-        }
-    }
 }
